+++ packages/experimental/lib.esm/index.d.ts
@@ -1,4 +1,5 @@
 import { BrainWallet } from "./brain-wallet";
 import { NonceManager } from "./nonce-manager";
-export { BrainWallet, NonceManager };
+import { Eip1193Bridge } from "./eip1193-bridge";
+export { BrainWallet, Eip1193Bridge, NonceManager };
 //# sourceMappingURL=index.d.ts.map