+++ packages/experimental/lib.esm/index.js
@@ -1,5 +1,6 @@
 "use strict";
 import { BrainWallet } from "./brain-wallet";
 import { NonceManager } from "./nonce-manager";
-export { BrainWallet, NonceManager };
+import { Eip1193Bridge } from "./eip1193-bridge";
+export { BrainWallet, Eip1193Bridge, NonceManager };
 //# sourceMappingURL=index.js.map